healthcare-anomaly/requirements.txt
flask==2.3.2
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
werkzeug==2.3.6
kafka-python==2.0.2
psycopg2-binary==2.9.6
sqlalchemy==2.0.19
tensorflow==2.12.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.2
scipy==1.11.0
joblib==1.3.1
python-dotenv==1.0.0
schedule==1.2.0
python-dateutil==2.8.2
requests==2.31.0
matplotlib==3.7.2
gunicorn==21.2.0
